- prefixed all module files with dcf-, which means that for all modules this point onwards must specify the export variable name to be hooked up properly in the build-utils/umd-tasks-common.js and build-utils/umd-tasks-optional.js -- previously gulp-umd will automatically take the file name and capitalize it as the return variable inside the UMD wrapper but with the dcf- prefix, that messes it up - for notice widget, check to see if element has an id before setting localStorage - changed error message to info if no id is detected when setting notice to fixedBottom - changed notice id prefix from unl-widget-fixedBottom-- to dcf-widget-fixedBottom-- - added title and message checking for Notice.create() - added ability to add id in widgetOptions for Notice.create()
