+++ build-utils/umd-tasks-common.js
@@ -14,7 +14,7 @@ const customPlumber = require('./custom-plumber');
  */
 
 const noticeWidget = [
-	gulp.src(path.join(buildPaths.appJsCommonSrc, 'notice.js')),
+	gulp.src(path.join(buildPaths.appJsCommonSrc, 'dcf-notice.js')),
 	customPlumber('error wrapping noticeWidget'),
 	$.debug({title: 'All Files - [UMDnoticeWidget:newer]'}), // uncomment to see src files
 	$.newer({ dest:buildPaths.umdCommonAppDest }),
@@ -24,14 +24,15 @@ const noticeWidget = [
 			return [
 				{
 					name: 'uuidGenerator',
-					amd: './uuid-gen',
-					cjs: './uuid-gen',
+					amd: './dcf-uuidGen',
+					cjs: './dcf-uuidGen',
 					global: 'dcfHelperUuidv4', // how this dependency is defined in the global scope
 					param: 'uuidv4' // how this dependency is called inside the module
 				}
 			]
 		},
-		namespace: () => 'dcfWidgetNotice'
+		exports: () => 'Notice', // the variable name that will be returned
+		namespace: () => 'dcfWidgetNotice' // how this module is named in the global scope
 		// if no exports defined, capitalized filename will be used instead
 	}),
 	gulp.dest(buildPaths.umdCommonAppDest)
@@ -39,7 +40,7 @@ const noticeWidget = [
 
 
 const uuidGen = [
-		gulp.src(path.join(buildPaths.appJsCommonSrc, 'uuid-gen.js')),
+		gulp.src(path.join(buildPaths.appJsCommonSrc, 'dcf-uuidGen.js')),
 		customPlumber('error wrapping uuidGen'),
 	$.debug({title: 'All Files - [UMDuuidGen:newer]'}), // uncomment to see src files
 	$.newer({ dest:buildPaths.umdCommonAppDest }),
@@ -53,24 +54,26 @@ const uuidGen = [
 
 
 const dialog = [
-	gulp.src(path.join(buildPaths.appJsCommonSrc, 'dialog.js')),
+	gulp.src(path.join(buildPaths.appJsCommonSrc, 'dcf-dialog.js')),
 	customPlumber('error wrapping dialog'),
 	$.debug({title: 'All Files - [UMDdialog:newer]'}), // uncomment to see src files
 	$.newer({ dest:buildPaths.umdCommonAppDest }),
 	$.debug({title: 'Passed Through - [UMDdialog:newer]'}), // uncomment to see passed
 	$.umd({
+		exports: () => 'Dialog', // TODO: needs to be redefined when file is modularized
 		namespace: () => 'dcfDialog'
 	}),
 	gulp.dest(buildPaths.umdCommonAppDest)
 ];
 
 const lazyLoad = [
-	gulp.src(path.join(buildPaths.appJsCommonSrc, 'lazy-load.js')),
+	gulp.src(path.join(buildPaths.appJsCommonSrc, 'dcf-lazyLoad.js')),
 	customPlumber('error wrapping lazyLoad'),
 	$.debug({title: 'All Files - [UMDlazyload:newer]'}), // uncomment to see src files
 	$.newer({ dest:buildPaths.umdCommonAppDest }),
 	$.debug({title: 'Passed Through - [UMDlazyload:newer]'}), // uncomment to see passed
 	$.umd({
+		exports: () => 'lazyLoad', // TODO: needs to be redefined when file is modularized
 		namespace: () => 'dcfLazyLoad'
 	}),
 	gulp.dest(buildPaths.umdCommonAppDest)
